requirements.txt
fastapi
uvicorn[standard]
google-cloud-firestore
python-dotenv
google-generativeai
openai